Structural changes and error issues
This prevents errors in the case where 127.0.0.1 is returned as a result of the DNS query.  This can happen if DNS blocking is in place.  The DNS server will block the host by returning a value of 127.0.0.1 .  If this is injected into the routing table, it will break.

Alternatively, we've stopped the routing of any private or reserved address, since this probably shouldn't be going out the gateway.